# daemon/Program.cs
using System.Net.WebSockets;
using System.Text;
using System.Text.Json;
using Clipwell.Daemon;
using Clipwell.Daemon.Mcp;
using Clipwell.Protocol;
using Microsoft.Extensions.FileProviders;

var builder = WebApplication.CreateBuilder(args);

// Clipwell's own default port (8787), overridable via CLIPWELL_URL. Avoids the
// crowded 5000/5001 range that dev servers squat on.
builder.WebHost.UseUrls(Environment.GetEnvironmentVariable("CLIPWELL_URL") ?? "http://127.0.0.1:8787");

builder.Services.ConfigureHttpJsonOptions(o =>
{
    o.SerializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
});

builder.Services.AddOpenApi("v1");

// Loopback-only CORS so the web UI (Tauri webview, vite dev, or any local browser)
// can fetch the JSON API cross-origin. Same-origin (the daemon serving /app) needs
// no CORS; this covers tauri://localhost, http(s)://tauri.localhost, and localhost:*.
builder.Services.AddCors(o => o.AddPolicy("webui", p => p
    .SetIsOriginAllowed(IsLocalOrigin)
    .AllowAnyHeader()
    .AllowAnyMethod()));

// One resolved data dir for the whole process: CLIPWELL_DATA_DIR, else the
// per-OS app-data folder. Passed in explicitly rather than re-read per store.
var dataDir = DataPaths.Resolve();
builder.Services.AddSingleton(_ => new MetadataStore(dataDir));
builder.Services.AddSingleton(sp => new HistoryStore(sp.GetRequiredService<MetadataStore>(), dataDir));
builder.Services.AddSingleton<ClipboardHub>();
builder.Services.AddSingleton<IClipboardWatcher>(sp =>
    ClipboardWatcherFactory.Create(sp.GetRequiredService<HistoryStore>().CacheDir));

// MCP over HTTP/SSE, served in-process at /mcp (tools hit HistoryStore directly).
// The stdio server in mcp/ remains for clients that spawn a child process.
builder.Services
    .AddMcpServer()
    .WithHttpTransport()
    .WithTools<DaemonClipboardTools>();

var app = builder.Build();
app.UseWebSockets();
app.UseCors("webui");

// Serve the built web UI at /app so the same SPA opens in any browser at
// http://127.0.0.1:8787/app. Resolve from: CLIPWELL_WEBUI_DIR, the published
// wwwroot/app, or the dev build at <repo>/webui/dist. Only mounted if found.
var webuiCandidates = new[]
{
    Environment.GetEnvironmentVariable("CLIPWELL_WEBUI_DIR"),
    Path.Combine(AppContext.BaseDirectory, "wwwroot", "app"),
    Path.Combine(AppContext.BaseDirectory, "..", "..", "..", "..", "webui", "dist"),
};
var webuiDir = webuiCandidates.FirstOrDefault(d => !string.IsNullOrEmpty(d) && Directory.Exists(d));
if (webuiDir is not null)
{
    var fp = new PhysicalFileProvider(Path.GetFullPath(webuiDir));
    // Redirect exactly "/app" → "/app/" (a middleware, not a route, so it doesn't
    // also capture "/app/" and loop), then serve the SPA's default + static files.
    app.Use(async (ctx, next) =>
    {
        if (ctx.Request.Path == "/app")
        {
            ctx.Response.Redirect("/app/");
            return;
        }
        await next();
    });
    app.UseDefaultFiles(new DefaultFilesOptions { FileProvider = fp, RequestPath = "/app" });
    app.UseStaticFiles(new StaticFileOptions { FileProvider = fp, RequestPath = "/app" });
}

app.MapOpenApi("/openapi/v1.json"); // machine-readable API spec (feeds the docs site)
app.MapMcp("/mcp"); // Streamable HTTP + SSE MCP endpoint

var store = app.Services.GetRequiredService<HistoryStore>();
var meta = app.Services.GetRequiredService<MetadataStore>();
var hub = app.Services.GetRequiredService<ClipboardHub>();
var watcher = app.Services.GetRequiredService<IClipboardWatcher>();

app.Logger.LogInformation(
    "CLIPWELL_DATA_DIR env = {Env}; resolved db = {Db}",
    Environment.GetEnvironmentVariable("CLIPWELL_DATA_DIR") ?? "(null)", store.DbPath);
var jsonOpts = new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase };

watcher.Failed += message => app.Logger.LogError("clipboard watcher: {Message}", message);
store.DetectorFailed += (detectorId, error) =>
    app.Logger.LogError(error, "detector {DetectorId} threw while classifying; skipped", detectorId);

// Capture pipeline: clipboard change → persist → broadcast to live subscribers.
watcher.Changed += row =>
{
    try
    {
        var isNew = store.Upsert(row);
        if (!isNew) return;
        app.Logger.LogInformation("captured clipboard item ({Len} chars)", row.TextLength);
        var payload = JsonSerializer.Serialize(
            new { type = "clipboard.changed", timestamp = row.Timestamp, textLength = row.TextLength },
            jsonOpts);
        hub.Broadcast(payload);
    }
    catch (Exception ex)
    {
        app.Logger.LogError(ex, "capture pipeline failed");
    }
};
watcher.Start();

// Hourly retention sweep. The delay comes first so a short-lived run (e.g. a dev
// session killed within minutes) never triggers a destructive purge — only a
// daemon that stays up for an hour actually sweeps. Set CLIPWELL_NO_SWEEP to
// disable purging entirely (useful when pointed at a shared/real DB).
var sweepDisabled = Environment.GetEnvironmentVariable("CLIPWELL_NO_SWEEP") is not null;
if (!sweepDisabled)
{
    var stopping = app.Lifetime.ApplicationStopping;
    _ = Task.Run(async () =>
    {
        while (!stopping.IsCancellationRequested)
        {
            try
            {
                await Task.Delay(TimeSpan.FromHours(1), stopping);
            }
            catch (OperationCanceledException)
            {
                return; // shutting down
            }
            try
            {
                var deleted = store.SweepOlderThan(store.LoadSettings().RetentionDays);
                if (deleted > 0) app.Logger.LogInformation("retention sweep purged {Count} rows", deleted);
            }
            catch (Exception ex)
            {
                app.Logger.LogError(ex, "retention sweep failed");
            }
        }
    });
}

app.Lifetime.ApplicationStopping.Register(() => watcher.Dispose());

// ── REST ────────────────────────────────────────────────────────────────

app.MapGet("/health", () => Results.Ok(new { status = "ok", db = store.DbPath, subscribers = hub.SubscriberCount }))
    .WithName("Health").WithSummary("Liveness probe and basic daemon info.");

app.MapGet("/api/clipboard", (int? limit, string? before, bool? excludeSensitive) =>
{
    var items = store.QueryPage(Math.Clamp(limit ?? 200, 1, 1000), before, excludeSensitive == true);
    return Results.Ok(new { items });
})
    .WithName("GetHistory")
    .WithSummary("List clipboard history, newest first. Page back with `before` (a timestamp).");

app.MapGet("/api/clipboard/counts", (string? q) => Results.Ok(store.GetCounts(q)))
    .WithName("GetCounts")
    .WithSummary("Aggregate counts (total, pinned, sensitive, per kind), optionally scoped to a search query `q`.");

app.MapGet("/api/clipboard/search", (string? q, int? limit, bool? excludeSensitive) =>
{
    if (string.IsNullOrWhiteSpace(q)) return Results.BadRequest(new { error = "q required" });
    return Results.Ok(new { items = store.Search(q, Math.Clamp(limit ?? 50, 1, 1000), excludeSensitive == true) });
})
    .WithName("SearchHistory")
    .WithSummary("Search the whole history by text or alias (case-insensitive), newest first.");

app.MapGet("/api/clipboard/item/{timestamp}", (string timestamp) =>
    store.FindByTimestamp(timestamp) is { } item ? Results.Ok(item) : Results.NotFound())
    .WithName("GetItem").WithSummary("Fetch a single history item by its exact timestamp.");

app.MapGet("/api/clipboard/settings", () => Results.Ok(store.LoadSettings()))
    .WithName("GetSettings").WithSummary("Get daemon settings (retention).");

app.MapPost("/api/clipboard/settings", (ClipboardSettings settings) =>
{
    // Reject an out-of-range retention rather than persisting a value LoadSettings
    // will silently discard on the next read.
    if (!ClipboardSettings.ValidRetentions.Contains(settings.RetentionDays))
        return Results.BadRequest(new { error = "retentionDays must be 7, 30, 90, or null" });
    store.SaveSettings(settings);
    return Results.Ok(settings);
})
    .WithName("SaveSettings").WithSummary("Update daemon settings (retention).");

app.MapPost("/api/clipboard/delete", (DeleteRequest req) =>
{
    if (string.IsNullOrEmpty(req.Timestamp)) return Results.BadRequest(new { error = "timestamp required" });
    return Results.Ok(new { deleted = store.DeleteByTimestamp(req.Timestamp) });
})
    .WithName("DeleteItem").WithSummary("Delete one history item by its timestamp.");

app.MapPost("/api/clipboard/clear", () => Results.Ok(new { deleted = store.ClearAll() }))
    .WithName("ClearHistory").WithSummary("Delete all history.");

// Dev-only seeding for the docs-capture scripts: insert items directly so capture
// never has to touch the user's real clipboard. Gated by CLIPWELL_ALLOW_SEED.
if (Environment.GetEnvironmentVariable("CLIPWELL_ALLOW_SEED") == "1")
{
    app.MapPost("/api/clipboard/_seed", (SeedRequest req) =>
    {
        store.Upsert(new StoreRow
        {
            Timestamp = string.IsNullOrEmpty(req.Timestamp) ? DateTimeOffset.UtcNow.ToString("o") : req.Timestamp,
            TextContent = req.Text,
            TextLength = req.Text?.Length ?? 0,
            HasImage = req.HasImage,
            ImagePath = req.ImagePath,
            SourceApp = req.SourceApp,
            Formats = req.HasImage ? ["image"] : req.Text is not null ? ["text"] : [],
        });
        return Results.Ok();
    });
}

app.MapPost("/api/clipboard/pin", (PinRequest req) =>
{
    if (string.IsNullOrEmpty(req.Timestamp)) return Results.BadRequest(new { error = "timestamp required" });
    meta.SetPinned(req.Timestamp, req.Pinned);
    return Results.Ok(new { req.Timestamp, req.Pinned });
})
    .WithName("PinItem").WithSummary("Pin or unpin an item (kept across retention).");

app.MapPost("/api/clipboard/sensitive", (SensitiveRequest req) =>
{
    if (string.IsNullOrEmpty(req.Timestamp)) return Results.BadRequest(new { error = "timestamp required" });
    meta.SetSensitive(req.Timestamp, req.Sensitive);
    return Results.Ok(new { req.Timestamp, req.Sensitive });
})
    .WithName("MarkSensitive").WithSummary("Mark or unmark an item as sensitive (masked in the UI).");

app.MapPost("/api/clipboard/rename", (RenameRequest req) =>
{
    if (string.IsNullOrEmpty(req.Timestamp)) return Results.BadRequest(new { error = "timestamp required" });
    meta.SetAlias(req.Timestamp, req.Alias);
    return Results.Ok(new { req.Timestamp, req.Alias });
})
    .WithName("RenameItem").WithSummary("Set or clear a custom alias for an item.");

app.MapPost("/api/clipboard/edit", (EditRequest req) =>
{
    if (string.IsNullOrEmpty(req.Timestamp)) return Results.BadRequest(new { error = "timestamp required" });
    meta.SetEdit(req.Timestamp, req.Text);
    // Same event shape as a capture so every live client refreshes its view.
    hub.Broadcast(JsonSerializer.Serialize(
        new { type = "clipboard.changed", timestamp = req.Timestamp, textLength = req.Text?.Length ?? 0 },
        jsonOpts));
    return Results.Ok(new { req.Timestamp, edited = !string.IsNullOrEmpty(req.Text) });
})
    .WithName("EditItem")
    .WithSummary("Override an item's text (non-destructive; empty text restores the original capture).");

app.MapGet("/api/clipboard/image/{timestamp}", (string timestamp) =>
{
    var path = store.GetImagePath(timestamp);
    // Only ever serve out of the cache dir: image_path is data, and a history file
    // carrying a path to somewhere else must not turn this into a file reader.
    return path is not null && IsInside(store.CacheDir, path) && File.Exists(path)
        ? Results.File(path, "image/png")
        : Results.NotFound();
})
    .WithName("GetImage").WithSummary("Fetch the cached PNG for an image item, by timestamp.");

// ── SSE ─────────────────────────────────────────────────────────────────

app.MapGet("/api/clipboard/stream", async (HttpContext ctx) =>
{
    ctx.Response.Headers.ContentType = "text/event-stream";
    ctx.Response.Headers.CacheControl = "no-cache";
    var (id, reader) = hub.Subscribe();
    try
    {
        await foreach (var payload in reader.ReadAllAsync(ctx.RequestAborted))
        {
            await ctx.Response.WriteAsync($"data: {payload}\n\n", ctx.RequestAborted);
            await ctx.Response.Body.FlushAsync(ctx.RequestAborted);
        }
    }
    catch (OperationCanceledException)
    {
        // client disconnected
    }
    finally
    {
        hub.Unsubscribe(id);
    }
})
    .WithName("StreamSse").WithSummary("Server-Sent Events stream of clipboard.changed events.");

// ── WebSocket ─────────────────────────────────────────────────────────────

app.MapGet("/api/clipboard/ws", async (HttpContext ctx) =>
{
    if (!ctx.WebSockets.IsWebSocketRequest)
    {
        ctx.Response.StatusCode = StatusCodes.Status400BadRequest;
        return;
    }
    using var socket = await ctx.WebSockets.AcceptWebSocketAsync();
    var (id, reader) = hub.Subscribe();
    try
    {
        await foreach (var payload in reader.ReadAllAsync(ctx.RequestAborted))
        {
            var bytes = Encoding.UTF8.GetBytes(payload);
            await socket.SendAsync(bytes, WebSocketMessageType.Text, true, ctx.RequestAborted);
        }
    }
    catch (OperationCanceledException)
    {
        // client disconnected
    }
    finally
    {
        hub.Unsubscribe(id);
    }
});

app.Run();

// Allow only local web-UI origins: loopback (127.0.0.1 / ::1 / localhost, any port)
// and the Tauri webview origins.
static bool IsLocalOrigin(string origin)
{
    if (origin is "tauri://localhost" or "http://tauri.localhost" or "https://tauri.localhost")
        return true;
    // The scheme check matters: Uri.IsLoopback is also true for file:// URIs, which
    // would otherwise let any local HTML file read the whole clipboard history.
    return Uri.TryCreate(origin, UriKind.Absolute, out var parsed)
        && (parsed.Scheme == Uri.UriSchemeHttp || parsed.Scheme == Uri.UriSchemeHttps)
        && parsed.IsLoopback;
}

// True when <paramref name="path"/> resolves to a file inside <paramref name="root"/>.
static bool IsInside(string root, string path)
{
    try
    {
        var rootFull = Path.TrimEndingDirectorySeparator(Path.GetFullPath(root))
            + Path.DirectorySeparatorChar;
        return Path.GetFullPath(path).StartsWith(rootFull, StringComparison.Ordinal);
    }
    catch (Exception ex) when (ex is ArgumentException or NotSupportedException or PathTooLongException)
    {
        return false;
    }
}

/// <summary>Entry point marker so the test host can boot the real app via WebApplicationFactory.</summary>
public partial class Program;

internal sealed record SeedRequest(string? Timestamp, string? Text, bool HasImage, string? ImagePath, string? SourceApp);
internal sealed record DeleteRequest(string Timestamp);
internal sealed record PinRequest(string Timestamp, bool Pinned);
internal sealed record SensitiveRequest(string Timestamp, bool Sensitive);
internal sealed record RenameRequest(string Timestamp, string? Alias);
internal sealed record EditRequest(string Timestamp, string? Text);