# openapi/clipwell.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Clipwell.Daemon | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787/"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Liveness probe and basic daemon info.",
        "operationId": "Health",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard": {
      "get": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "List clipboard history, newest first. Page back with `before` (a timestamp).",
        "operationId": "GetHistory",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excludeSensitive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/counts": {
      "get": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Aggregate counts (total, pinned, sensitive, per kind), optionally scoped to a search query `q`.",
        "operationId": "GetCounts",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/search": {
      "get": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Search the whole history by text or alias (case-insensitive), newest first.",
        "operationId": "SearchHistory",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "excludeSensitive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/item/{timestamp}": {
      "get": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Fetch a single history item by its exact timestamp.",
        "operationId": "GetItem",
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/settings": {
      "get": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Get daemon settings (retention).",
        "operationId": "GetSettings",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Update daemon settings (retention).",
        "operationId": "SaveSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipboardSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/delete": {
      "post": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Delete one history item by its timestamp.",
        "operationId": "DeleteItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/clear": {
      "post": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Delete all history.",
        "operationId": "ClearHistory",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/pin": {
      "post": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Pin or unpin an item (kept across retention).",
        "operationId": "PinItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/sensitive": {
      "post": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Mark or unmark an item as sensitive (masked in the UI).",
        "operationId": "MarkSensitive",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensitiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/rename": {
      "post": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Set or clear a custom alias for an item.",
        "operationId": "RenameItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/edit": {
      "post": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Override an item's text (non-destructive; empty text restores the original capture).",
        "operationId": "EditItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clipboard/image/{timestamp}": {
      "get": {
        "tags": [
          "Clipwell.Daemon"
        ],
        "summary": "Fetch the cached PNG for an image item, by timestamp.",
        "operationId": "GetImage",
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClipboardSettings": {
        "type": "object",
        "properties": {
          "retentionDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "openAtCursor": {
            "type": "boolean"
          },
          "theme": {
            "type": "string"
          },
          "defaultView": {
            "type": "string"
          },
          "defaultGroup": {
            "type": "string"
          },
          "showSource": {
            "type": "boolean"
          },
          "showTime": {
            "type": "boolean"
          },
          "showChars": {
            "type": "boolean"
          },
          "hotkey": {
            "type": "string"
          }
        }
      },
      "DeleteRequest": {
        "required": [
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          }
        }
      },
      "EditRequest": {
        "required": [
          "timestamp",
          "text"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "text": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PinRequest": {
        "required": [
          "timestamp",
          "pinned"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "pinned": {
            "type": "boolean"
          }
        }
      },
      "RenameRequest": {
        "required": [
          "timestamp",
          "alias"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "alias": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SensitiveRequest": {
        "required": [
          "timestamp",
          "sensitive"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "sensitive": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Clipwell.Daemon"
    }
  ]
}